v2 setup: new SDK, new API package, no native part yet
