+++ src/main/scala/org/bblfsh/client/cli/ScalaClientCLI.scala
@@ -1,6 +1,7 @@
 package org.bblfsh.client.cli
 
-import org.bblfsh.client.BblfshClient
+import org.bblfsh.client.v2.BblfshClient
+import org.bblfsh.client.v2.BblfshClient._
 
 import scala.io.Source
 import scala.util.control.NonFatal
@@ -24,12 +25,12 @@ object ScalaClientCLI extends App {
   if (resp.errors.isEmpty) {
     if (query != null && query != "") {
       try {
-        println(BblfshClient.filter(resp.uast.get, query))
+        println(BblfshClient.filter(resp.uast.decode().root(), query))
       } catch {
         case NonFatal(e) => print(f"Filter error: $e")
       }
     } else {
-      println(resp.uast.get)
+      println(resp.uast.decode())
     }
   } else {
     println(s"Parsing failed with ${resp.errors.length} errors:")
