+++ src/main/scala/org/bblfsh/client/v2/Node.scala
@@ -1,0 +1,5 @@
+package org.bblfsh.client.v2
+	
+case class Node(ctx: Long, handle: Long) {
+  @native def load(): Map[String, _]
+}
